+++ RandomSkunk.Results/CompositeError.cs
@@ -5,10 +5,23 @@ namespace RandomSkunk.Results;
 /// </summary>
 public record class CompositeError : Error
 {
-    private CompositeError(IReadOnlyList<Error> errors, string? messageDetail = null)
-        : base((nameof(Errors), errors))
+    private static readonly string _errorsFieldFullName =
+        $"{GetTypeFullName(typeof(CompositeError))}.{nameof(Errors)}";
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="CompositeError"/> class.
+    /// </summary>
+    /// <param name="errors">The errors that make up this instance of <see cref="CompositeError"/>.</param>
+    /// <param name="messageDetail">The message of the composite error.</param>
+    /// <exception cref="ArgumentNullException">If <paramref name="errors"/> is <see langword="null"/>.</exception>
+    /// <exception cref="ArgumentOutOfRangeException">If <paramref name="errors"/> does not contain at least two errors.</exception>
+    public CompositeError(IEnumerable<Error> errors, string? messageDetail = null)
+        : base((_errorsFieldFullName, errors?.ToList().AsReadOnly() ?? throw new ArgumentNullException(nameof(errors))))
     {
-        var defaultMessage = $"{GetNumberName(errors.Count)} errors occurred. See '{nameof(Errors)}' item under Extensions property for details.";
+        if (Errors.Count < 2)
+            throw new ArgumentException("Sequence must contain at least two errors.", nameof(errors));
+
+        var defaultMessage = $"{GetNumberName(Errors.Count)} errors occurred.";
         var message = string.IsNullOrEmpty(messageDetail) ? defaultMessage : defaultMessage + ' ' + messageDetail;
         Message = message;
     }
@@ -17,39 +30,7 @@ public record class CompositeError : Error
     /// Gets the errors that make up this instance of <see cref="CompositeError"/>.
     /// </summary>
     public IReadOnlyList<Error> Errors =>
-        TryGet<IReadOnlyList<Error>>(nameof(Errors), out var errors)
-            ? errors
-            : [];
-
-    /// <summary>
-    /// Creates a composite error from the specified sequence of two or more errors.
-    /// </summary>
-    /// <param name="errors">A sequence of two or more errors.</param>
-    /// <param name="messageDetail">The message of the composite error.</param>
-    /// <param name="errorCode">The error code of the composite error.</param>
-    /// <param name="identifier">The identifier of the composite error.</param>
-    /// <returns>A <see cref="CompositeError"/> consisting of the specified errors.</returns>
-    /// <exception cref="ArgumentNullException">If <paramref name="errors"/> is <see langword="null"/>.</exception>
-    /// <exception cref="ArgumentOutOfRangeException">If <paramref name="errors"/> does not contain at least two errors.</exception>
-    public static CompositeError Create(
-        IEnumerable<Error> errors,
-        string? messageDetail = null,
-        int? errorCode = null,
-        string? identifier = null)
-    {
-        if (errors is null) throw new ArgumentNullException(nameof(errors));
-
-        var errorList = errors.ToList();
-
-        if (errorList.Count < 2)
-            throw new ArgumentException("Sequence must contain at least two errors.", nameof(errors));
-
-        return new CompositeError(errorList, messageDetail)
-        {
-            ErrorCode = errorCode,
-            Identifier = identifier,
-        };
-    }
+        TryGet<IReadOnlyList<Error>>(_errorsFieldFullName, out var errors) ? errors : [];
 
     /// <summary>
     /// Creates a composite error from the specified non-empty sequence of errors if it contains more than one error, otherwise