+++ RandomSkunk.Results/Result.cs
@@ -190,11 +190,8 @@ public readonly struct Result : IEquatable<Result>
                 if (_outcome == Outcome.Success)
                     return error;
 
-                return CompositeError.Create(
-                    [
-                        GetError(),
-                        error,
-                    ],
+                return new CompositeError(
+                    [GetError(), error],
                     $"The first error is the original error; the second error is from the TaskCanceledException thrown when evaluating the '{nameof(callback)}' function parameter.");
             }
             catch (Exception ex)
@@ -204,11 +201,8 @@ public readonly struct Result : IEquatable<Result>
                 if (_outcome == Outcome.Success)
                     return error;
 
-                return CompositeError.Create(
-                    [
-                        GetError(),
-                        error,
-                    ],
+                return new CompositeError(
+                    [GetError(), error],
                     $"The first error is the original error; the second error is from the Exception thrown when evaluating the '{nameof(callback)}' function parameter.");
             }
         }
@@ -243,11 +237,8 @@ public readonly struct Result : IEquatable<Result>
                 if (_outcome == Outcome.Success)
                     return error;
 
-                return CompositeError.Create(
-                    [
-                        GetError(),
-                        error,
-                    ],
+                return new CompositeError(
+                    [GetError(), error],
                     $"The first error is the original error; the second error is from the TaskCanceledException thrown when evaluating the '{nameof(callback)}' function parameter.");
             }
             catch (Exception ex)
@@ -257,11 +248,8 @@ public readonly struct Result : IEquatable<Result>
                 if (_outcome == Outcome.Success)
                     return error;
 
-                return CompositeError.Create(
-                    [
-                        GetError(),
-                        error,
-                    ],
+                return new CompositeError(
+                    [GetError(), error],
                     $"The first error is the original error; the second error is from the Exception thrown when evaluating the '{nameof(callback)}' function parameter.");
             }
         }
@@ -339,20 +327,14 @@ public readonly struct Result : IEquatable<Result>
                 }
                 catch (TaskCanceledException ex)
                 {
-                    return CompositeError.Create(
-                        [
-                            GetError(),
-                            Errors.Canceled(ex),
-                        ],
+                    return new CompositeError(
+                        [GetError(), Errors.Canceled(ex)],
                         $"The first error is the original error; the second error is from the TaskCanceledException thrown when evaluating the '{nameof(onFailCallback)}' function parameter.");
                 }
                 catch (Exception ex)
                 {
-                    return CompositeError.Create(
-                        [
-                            GetError(),
-                            Error.FromExceptionThrownInCallback(ex, nameof(onFailCallback)),
-                        ],
+                    return new CompositeError(
+                        [GetError(), Error.FromExceptionThrownInCallback(ex, nameof(onFailCallback))],
                         $"The first error is the original error; the second error is from the Exception thrown when evaluating the '{nameof(onFailCallback)}' function parameter.");
                 }
             }
@@ -384,20 +366,14 @@ public readonly struct Result : IEquatable<Result>
                 }
                 catch (TaskCanceledException ex)
                 {
-                    return CompositeError.Create(
-                        [
-                            GetError(),
-                            Errors.Canceled(ex),
-                        ],
+                    return new CompositeError(
+                        [GetError(), Errors.Canceled(ex)],
                         $"The first error is the original error; the second error is from the TaskCanceledException thrown when evaluating the '{nameof(onFailCallback)}' function parameter.");
                 }
                 catch (Exception ex)
                 {
-                    return CompositeError.Create(
-                        [
-                            GetError(),
-                            Error.FromExceptionThrownInCallback(ex, nameof(onFailCallback)),
-                        ],
+                    return new CompositeError(
+                        [GetError(), Error.FromExceptionThrownInCallback(ex, nameof(onFailCallback))],
                         $"The first error is the original error; the second error is from the Exception thrown when evaluating the '{nameof(onFailCallback)}' function parameter.");
                 }
             }